feat: remove migration for P9130+P9135
