+++ runtime/litentry/src/migration/migration.md
@@ -4,3 +4,10 @@ https://github.com/litentry/litentry-parachain/releases/tag/v0.9.10
 
  <!-- MigrateCollatorSelectionIntoParachainStaking -->
  # The main purpose of runtime upgrade is for make up the missing genesis build of ParachainStaking and clean the old CollatorSelection storage.
+
+ P9130.rs
+ https://github.com/litentry/litentry-parachain/releases/tag/v0.9.13-1
+  # This migration is for the update of AtStaked with ParachainStaking
+
+  <!-- MigrateAtStakeAutoCompound -->
+  # The main purpose of runtime upgrade is for adding the autocompound staking function of ParachainStaking and need to update storage to the latest struct.
